culinary_recipes/requirements.txt
asgiref==3.5.2
certifi==2022.9.24
charset-normalizer==2.1.1
cloudinary==1.30.0
Django==4.1.3
django-anymail==8.6
django-embed-video==1.4.7
gunicorn==20.1.0
idna==3.4
Pillow==9.3.0
psycopg2-binary==2.9.5
requests==2.28.1
six==1.16.0
sqlparse==0.4.3
urllib3==1.26.12
